+++ x-pack/plugins/aiops/public/components/log_categorization/document_count_chart.tsx
@@ -9,11 +9,15 @@ import React, { FC, useMemo } from 'react';
 
 import { i18n } from '@kbn/i18n';
 import { DocumentCountChart as DocumentCountChartRoot } from '@kbn/aiops-components';
+
+import type { Category, SparkLinesPerCategory } from '../../../common/api/log_categorization/types';
+
 import { useAiopsAppContext } from '../../hooks/use_aiops_app_context';
-import { TotalCountHeader } from '../document_count_content/total_count_header';
-import type { Category, SparkLinesPerCategory } from './use_categorize_request';
-import type { EventRate } from './use_categorize_request';
 import { DocumentCountStats } from '../../get_document_stats';
+
+import { TotalCountHeader } from '../document_count_content/total_count_header';
+
+import type { EventRate } from './use_categorize_request';
 
 interface Props {
   totalCount: number;
